+++ bin/console
@@ -4,4 +4,4 @@ require "bundler/setup"
 require "./lib/harvest_notifier"
 
 require "irb"
-IRB.start(__FILE__)
+IRB.start
